feat(nvim): migrate to lualine
heirline.nvim is very complex to configure and atm does not provide any
benefit over lualine. Also, since heirline is rather a statusline coding
kit and one needs to take care of a lot of things, it introduces way too
much friction. For example, I needed to implement a hack to get instant
mode updates, which is not what I want nor need.